bot/requirements.txt
aiogram==3.14.0
replicate==1.0.7
python-dotenv==1.0.1
aiohttp==3.10.11
aiosqlite==0.20.0
